Add Client interface
Added the client interface which contains two methods:
connect and disconnect.

There is also an implementation added, called Simple.
To create a new intance of Simple, the Factory class
should be used.

The configuration for the client will be added in a
later stage as well as functionality to send requests.